+++ app/src/main/java/woowacourse/omok/presentation/BoardView.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok
+package woowacourse.omok.presentation
 
 import android.widget.ImageView
 import android.widget.TableLayout
