+++ app/src/main/java/woowacourse/omok/presentation/StartActivity.kt
@@ -1,10 +1,11 @@
-package woowacourse.omok
+package woowacourse.omok.presentation
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
-import woowacourse.omok.gameactivity.GameActivity
+import woowacourse.omok.R
+import woowacourse.omok.presentation.gameactivity.GameActivity
 import woowacourse.omok.util.setOnSingleClickListener
 
 class StartActivity : AppCompatActivity() {
