+++ app/src/main/java/woowacourse/omok/presentation/gameactivity/MessageView.kt
@@ -1,4 +1,4 @@
-package woowacourse.omok.gameactivity
+package woowacourse.omok.presentation.gameactivity
 
 import android.view.View
 import android.widget.TextView
